requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
pyodbc>=5.0.1
python-dotenv>=1.0.0
Werkzeug>=2.3.7
